[integrity] Implement fuse idea.
But without an actual fuse node: We just ignore data flows through
unused parts of the graph.

What is used is determined by the actual inputs sent to an include node.
This is particularly imporant when we hook up include via * like for
ReAct tools.

This significantly reduces false negatives.